Refactor: Home screen and viewmodel
