+++ shared/src/commonMain/kotlin/ui/screen/home/HomeState.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package ui
+package ui.screen.home
 
 import androidx.compose.runtime.Immutable
 import dev.shreyaspatil.mutekt.core.annotations.GenerateMutableModel
